Exception Handling

+++ src/main/java/com/faccaogames/soccercoach/service/PlayerService.java
@@ -1,41 +1,79 @@
 package com.faccaogames.soccercoach.service;
 
+import com.faccaogames.soccercoach.exception.ApiRequestException;
 import com.faccaogames.soccercoach.model.Player;
+import com.faccaogames.soccercoach.model.Team;
 import com.faccaogames.soccercoach.repository.PlayerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDate;
 import java.util.List;
 
 @Service
 public class PlayerService {
 
     private final PlayerRepository playerRepository;
+    private final TeamService teamService;
 
     @Autowired
-    public PlayerService(PlayerRepository playerRepository) {
+    public PlayerService(PlayerRepository playerRepository, TeamService teamService) {
         this.playerRepository = playerRepository;
+        this.teamService = teamService;
     }
 
-    public void createPlayer(Player player) {
-        playerRepository.save(player);
+    public Long createPlayer(Player player) {
+        return playerRepository.save(player).getId();
     }
 
     public List<Player> retrieveAllPlayers() {
-        return playerRepository.findAll();
+        if (playerRepository.count() > 0) {
+            return playerRepository.findAll();
+        } else {
+            throw new ApiRequestException("No players were found.");
+        }
     }
 
     public Player retrievePlayerById(Long id) {
-        return playerRepository.getById(id);
+        if (playerRepository.existsById(id)) {
+            return playerRepository.getById(id);
+        } else {
+            throw new ApiRequestException("Player with id " + id + " not found.");
+        }
     }
 
-    public void deletePlayer(Long id) {
-        playerRepository.deleteById(id);
+    public String deletePlayer(Long id) {
+        if (playerRepository.existsById(id)) {
+            playerRepository.deleteById(id);
+            return "Player " + id + " was deleted.";
+        } else {
+            throw new ApiRequestException("Player with id " + id + " not found.");
+        }
     }
 
-    public void updatePlayer(Long id, Player player) {
-        player.setId(id);
-        playerRepository.save(player);
+    public Long updatePlayer(Long id, Player player) {
+        if (playerRepository.existsById(id)) {
+            player.setId(id);
+            return playerRepository.save(player).getId();
+        } else {
+            throw new ApiRequestException("Player with id " + id + " not found.");
+        }
+    }
+
+    public void transferPlayer(Long id, long teamId) {
+        Player player = playerRepository.getById(id);
+        Team currentTeam = teamService.retrieveTeamById(player.getTeamId());
+        Team team = teamService.retrieveTeamById(teamId);
+
+        doTransfer(player, teamId);
+        updateTeams(currentTeam, team, player);
+    }
+
+    private void updateTeams(Team currentTeam, Team team, Player player) {
+
+    }
+
+    private void doTransfer(Player player, long teamId) {
+        player.setTeamId(teamId);
+        updatePlayer(player.getId(), player);
     }
 }
